+++ packages/odyssey-react-mui/src/labs/DataFilters.tsx
@@ -20,6 +20,7 @@ import {
   useRef,
   useState,
 } from "react";
+import styled from "@emotion/styled";
 import { Autocomplete } from "../Autocomplete";
 import { Box } from "../Box";
 import { TagList } from "../TagList";
@@ -175,21 +176,59 @@ const DataFilters = ({
     }
   }, [onChangeSearch, searchValue, searchDelayTime, hasSearchSubmitButton]);
 
-  const handleInputChange = useCallback(
-    (filterId: string, value: DataFilterValue, submit: boolean = false) => {
+  const autocompleteOptions = useMemo(() => {
+    // Check if filterPopoverCurrentFilter and filterPopoverCurrentFilter.options are defined
+    if (
+      filterPopoverCurrentFilter?.variant === "autocomplete" &&
+      filterPopoverCurrentFilter?.options
+    ) {
+      return filterPopoverCurrentFilter.options.map((option) => ({
+        label: option.label,
+      }));
+    }
+
+    // if filterPopoverCurrentFilter or filterPopoverCurrentFilter.options is undefined
+    return [];
+  }, [filterPopoverCurrentFilter]);
+
+  const updateInputValue = useCallback(
+    ({ filterId, value }: { filterId: string; value: DataFilterValue }) => {
       setInputValues({ ...inputValues, [filterId]: value });
+    },
+    [inputValues]
+  );
 
-      if (submit) {
-        const updatedFilters = filtersProp.map((filter) => ({
-          ...filter,
-          value: filter.id === filterId ? value : inputValues[filter.id],
-        }));
+  const updateFilters = useCallback(
+    ({ filterId, value }: { filterId: string; value: DataFilterValue }) => {
+      const updatedFilters = filtersProp.map((filter) => ({
+        ...filter,
+        value: filter.id === filterId ? value : inputValues[filter.id],
+      }));
 
-        setFilters(updatedFilters);
-      }
+      setFilters(updatedFilters);
     },
     [inputValues, filtersProp]
   );
+
+  const getAutoCompleteLabel = <
+    Value extends { label: string } | Array<{ label: string }>
+  >(
+    value: Value
+  ) => {
+    if (Array.isArray(value)) {
+      // Iterating to find the label
+      return value
+        .map((valueElement) => {
+          if (typeof valueElement === "string") {
+            return undefined;
+          }
+          return valueElement.label;
+        })
+        .filter((item): item is string => Boolean(item));
+    }
+
+    return value?.label;
+  };
 
   const handleMultiSelectChange = useCallback(
     (filterId: string, value: string, submit: boolean = false) => {
@@ -245,6 +284,15 @@ const DataFilters = ({
     setFilters(updatedFilters);
   }, [inputValues, filtersProp]);
 
+  const AutocompleteOuterContainer = styled.div`
+    display: flex;
+    gap: 2;
+    align-items: center;
+    alignitems: "flex-end";
+  `;
+  const AutocompleteInnerContainer = styled.div`
+    width: "100%";
+  `;
   const filterMenu = useMemo(
     () => (
       <>
@@ -367,23 +415,43 @@ const DataFilters = ({
                     {/* Autocomplete */}
                     {filterPopoverCurrentFilter?.variant === "autocomplete" &&
                       filterPopoverCurrentFilter?.options && (
-                        <Autocomplete
-                          label={filterPopoverCurrentFilter.label}
-                          value={
-                            (inputValues[
-                              filterPopoverCurrentFilter.id
-                            ] as string) ?? ""
-                          }
-                          onBlur={function ro() {}}
-                          onChange={function ro() {}}
-                          onFocus={function ro() {}}
-                          onInputChange={function ro() {}}
-                          options={filterPopoverCurrentFilter.options.map(
-                            (option: { label: string }) => ({
-                              label: option.label,
-                            })
-                          )}
-                        />
+                        <AutocompleteOuterContainer>
+                          <AutocompleteInnerContainer>
+                            <Autocomplete
+                              label={filterPopoverCurrentFilter.label}
+                              value={
+                                inputValues[filterPopoverCurrentFilter.id] ?? ""
+                              }
+                              onChange={(_, value) => {
+                                const label =
+                                  typeof value === "string"
+                                    ? getAutoCompleteLabel({ label: value })
+                                    : Array.isArray(value)
+                                    ? getAutoCompleteLabel(
+                                        value.map((item) =>
+                                          typeof item === "string"
+                                            ? { label: item }
+                                            : item
+                                        )
+                                      )
+                                    : value
+                                    ? getAutoCompleteLabel(value)
+                                    : "";
+
+                                updateInputValue({
+                                  filterId: filterPopoverCurrentFilter.id,
+                                  value: label,
+                                });
+                              }}
+                              options={autocompleteOptions}
+                            />
+                          </AutocompleteInnerContainer>
+                          <Button
+                            variant="primary"
+                            endIcon={<CheckIcon />}
+                            type="submit"
+                          />
+                        </AutocompleteOuterContainer>
                       )}
                     {/* Text or Number */}
                     {(filterPopoverCurrentFilter?.variant === "text" ||
@@ -410,10 +478,10 @@ const DataFilters = ({
                               ] as string) ?? ""
                             }
                             onChange={(ev) =>
-                              handleInputChange(
-                                filterPopoverCurrentFilter.id,
-                                ev.currentTarget.value
-                              )
+                              updateInputValue({
+                                filterId: filterPopoverCurrentFilter.id,
+                                value: ev.currentTarget.value,
+                              })
                             }
                             endAdornment={
                               inputValues[filterPopoverCurrentFilter.id] && (
@@ -421,11 +489,15 @@ const DataFilters = ({
                                   size="small"
                                   aria-label="Clear filter"
                                   onClick={() => {
-                                    handleInputChange(
-                                      filterPopoverCurrentFilter.id,
-                                      undefined,
-                                      true
-                                    );
+                                    updateInputValue({
+                                      filterId: filterPopoverCurrentFilter.id,
+                                      value: undefined,
+                                    });
+
+                                    updateFilters({
+                                      filterId: filterPopoverCurrentFilter.id,
+                                      value: undefined,
+                                    });
                                   }}
                                 >
                                   <CloseCircleFilledIcon />
@@ -480,13 +552,17 @@ const DataFilters = ({
                       filterPopoverCurrentFilter?.options && (
                         <RadioGroup
                           label={filterPopoverCurrentFilter.label}
-                          onChange={(_, value) =>
-                            handleInputChange(
-                              filterPopoverCurrentFilter.id,
+                          onChange={(_, value) => {
+                            updateInputValue({
+                              filterId: filterPopoverCurrentFilter.id,
                               value,
-                              true
-                            )
-                          }
+                            });
+
+                            updateFilters({
+                              filterId: filterPopoverCurrentFilter.id,
+                              value,
+                            });
+                          }}
                         >
                           <Radio
                             label="Any"
@@ -561,7 +637,7 @@ const DataFilters = ({
               <Button
                 variant="secondary"
                 label="Clear filters"
-                onClick={() => clearAllFilters()}
+                onClick={clearAllFilters}
               />
             </Box>
           )}
@@ -585,7 +661,17 @@ const DataFilters = ({
               <Tag
                 key={filter.label}
                 label={`${filter.label}: ${filter.value}`}
-                onRemove={() => handleInputChange(filter.id, undefined, true)}
+                onRemove={() => {
+                  updateInputValue({
+                    filterId: filter.id,
+                    value: undefined,
+                  });
+
+                  updateFilters({
+                    filterId: filter.id,
+                    value: undefined,
+                  });
+                }}
               />
             ))}
           </TagList>